Added a catch for redis on error
I found that it does no retry the connection if one does not catch the
redisClient.on('error'...) instead the app craches.

There is probably ways to handle this higher up in the chain, but i
like node-redis to handle it.

To repro this, just start a client using redis-throttle and kill the
redis server.